Lint changes and general app organization
