+++ acm_web_site/apps/business/information/serializers.py
@@ -1,12 +1,16 @@
-# -- coding: utf-8
-from __future__ import unicode_literals
-import base64
-
 from django.conf import settings
 
 from rest_framework import serializers
-from ..models import Member, POSITION_CHOICES
+from .models import (
+    Member,
+    Award,
+    POSITION_CHOICES
+)
 
+import base64
+
+
+# TODO - Move auxiliary functions in serializers to Utils
 
 class MemberSerializer(serializers.ModelSerializer):
     picture = serializers.SerializerMethodField(required=False)
@@ -36,3 +40,18 @@ class MemberSerializer(serializers.ModelSerializer):
 
     def create(self, validated_data):
         return Member.objects.create(**validated_data)
+
+
+class AwardSerializer(serializers.ModelSerializer):
+    picture = serializers.SerializerMethodField()
+
+    class Meta:
+        model = Award
+        fields = '__all__'
+
+    def get_picture(self, obj):
+        prefix = '/'.join(settings.MEDIA_ROOT.split('/')[:-1])
+        complete_path = prefix + obj.picture
+        with open(complete_path, "rb") as image_file:
+            str = base64.b64encode(image_file.read())
+        return str
