+++ acm_web_site/settings/common.py
@@ -85,7 +85,8 @@ DATABASES = {
 AUTH_PASSWORD_VALIDATORS = [
     {
         'NAME':
-            'django.contrib.auth.password_validation.UserAttributeSimilarityValidator',
+            'django.contrib.auth.password_validation.\
+            UserAttributeSimilarityValidator',
     },
     {
         'NAME':
